Fix scientific() on small positive numbers.
Checking a number is negative using `'-' in string_value` leads to confusion
because 1e-30 and -1e30 both contain a '-' but only one of them is negative.
This bug had found its way into the tests and the docstring.

Additionally, the removal of redundant leading '0's and '+'s from the exponent
would only kick in if both were present so that 1e20 would become 10⁺²⁰ instead
of just 10²⁰ and the insertion of negative exponents could lead to outputs such
as 3.00 x 10⁻⁺²⁰.