Move rounding logic to _quotient_and_remainder
The logic about rounding due to formatting has been moved to
_quotient_and_remainder because this is where we have all the logic regarding
the minimum unit and suppress units.

Also removed _carry function. Instead use the same logic than was used for
calculating the `secs` based on the remaining amount of days. It is done
now also for `usecs` based on the `secs` remaining.

Add another block of logic to check after rounding if any units should be
promoted to a higher unit, in case of a rounding up.